ensure codex e2e uses mini model
