+++ src/cli/shim/errors.ts
@@ -23,12 +23,6 @@ export class ShimError extends Error {
 	}
 }
 
-export class InvalidUsageError extends ShimError {
-	constructor(message: string) {
-		super("invalid-usage", message);
-	}
-}
-
 export class BlockedError extends ShimError {
 	constructor(message: string) {
 		super("blocked", message);
