+++ src/lib/agents/errors.ts
@@ -1,0 +1,7 @@
+import { ShimError } from "../../cli/shim/errors.js";
+
+export class InvalidUsageError extends ShimError {
+	constructor(message: string) {
+		super("invalid-usage", message);
+	}
+}
