+++ src/lib/targets/builtins/codex/e2e/expected/approval-auto-edit.stderr.txt
@@ -1,21 +1,21 @@
-OpenAI Codex v0.92.0 (research preview)
+OpenAI Codex v0.89.0 (research preview)
 --------
 workdir: /Users/joeroddy/Documents/dev/projects/open-source/omniagent
-model: gpt-5.2-codex
+model: gpt-5.1-codex-mini
 provider: openai
 approval: never
 sandbox: workspace-write [workdir, /tmp, $TMPDIR]
 reasoning effort: high
 reasoning summaries: auto
-session id: 019c0af3-56b4-7430-97ac-ca0c7d05b337
+session id: 019c14bd-b52b-7991-9a86-e5bfc24f0b3d
 --------
 user
 Output exactly: 5
 mcp startup: no servers
 
 thinking
-**Setting output limit to five**
+**Outputting exact response**
 codex
 5
 tokens used
-2,602
+2,245
