+++ src/lib/targets/builtins/codex/e2e/expected/approval-yolo.stderr.txt
@@ -1,21 +1,21 @@
-OpenAI Codex v0.92.0 (research preview)
+OpenAI Codex v0.89.0 (research preview)
 --------
 workdir: /Users/joeroddy/Documents/dev/projects/open-source/omniagent
-model: gpt-5.2-codex
+model: gpt-5.1-codex-mini
 provider: openai
 approval: never
 sandbox: danger-full-access
 reasoning effort: high
 reasoning summaries: auto
-session id: 019c0af3-6a86-7d73-bdff-aa705538a0a1
+session id: 019c14bd-c6c0-71a0-b119-4dabae3a8801
 --------
 user
 Output exactly: 5
 mcp startup: no servers
 
 thinking
-**Preparing exact output count**
+**Outputting simple response**
 codex
 5
 tokens used
-2,533
+2,180
