+++ src/lib/targets/builtins/codex/e2e/expected/auto-edit-alias.stderr.txt
@@ -1,23 +1,21 @@
-OpenAI Codex v0.92.0 (research preview)
+OpenAI Codex v0.89.0 (research preview)
 --------
 workdir: /Users/joeroddy/Documents/dev/projects/open-source/omniagent
-model: gpt-5.2-codex
+model: gpt-5.1-codex-mini
 provider: openai
 approval: never
 sandbox: workspace-write [workdir, /tmp, $TMPDIR]
 reasoning effort: high
 reasoning summaries: auto
-session id: 019c0af3-6250-7bf2-bc0f-cc56590a4041
+session id: 019c14bd-bc05-7ea2-b37f-eea3ed2a384a
 --------
 user
 Output exactly: 5
 mcp startup: no servers
 
 thinking
-**Providing exact output**
-
-I'll output the exact string "5" as requested.
+**Returning simple output**
 codex
 5
 tokens used
-2,217
+2,240
