+++ src/lib/targets/builtins/codex/e2e/expected/model.stderr.txt
@@ -1,4 +1,4 @@
-OpenAI Codex v0.92.0 (research preview)
+OpenAI Codex v0.89.0 (research preview)
 --------
 workdir: /Users/joeroddy/Documents/dev/projects/open-source/omniagent
 model: gpt-5.1-codex-mini
@@ -7,15 +7,15 @@ approval: never
 sandbox: workspace-write [workdir, /tmp, $TMPDIR]
 reasoning effort: high
 reasoning summaries: auto
-session id: 019c0af3-9a9f-7bc2-bbfb-88d5fcacd67b
+session id: 019c14bd-ffd0-7ed1-afc4-beb654c19613
 --------
 user
 Output exactly: 5
 mcp startup: no servers
 
 thinking
-**Preparing trivial output**
+**Determining command count**
 codex
 5
 tokens used
-2,235
+2,245
