+++ src/lib/targets/builtins/codex/e2e/expected/output-json.stdout.txt
@@ -1,5 +1,5 @@
-{"type":"thread.started","thread_id":"019c0af3-7b6d-7590-820d-86de6cff5dee"}
+{"type":"thread.started","thread_id":"019c14bd-d5ff-7af3-a0ba-de70bccfd684"}
 {"type":"turn.started"}
-{"type":"item.completed","item":{"id":"item_0","type":"reasoning","text":"**Enforcing exact output count**"}}
+{"type":"item.completed","item":{"id":"item_0","type":"reasoning","text":"**Returning exact number**"}}
 {"type":"item.completed","item":{"id":"item_1","type":"agent_message","text":"5"}}
-{"type":"turn.completed","usage":{"input_tokens":9238,"cached_input_tokens":6656,"output_tokens":20}}
+{"type":"turn.completed","usage":{"input_tokens":9128,"cached_input_tokens":6912,"output_tokens":27}}
