+++ src/lib/targets/builtins/codex/e2e/expected/output-stream-json.stdout.txt
@@ -1,5 +1,5 @@
-{"type":"thread.started","thread_id":"019c0af3-8bcf-7280-9c32-5597ffc5d1f7"}
+{"type":"thread.started","thread_id":"019c14bd-eaec-7b92-8d7c-d21e973fbc40"}
 {"type":"turn.started"}
-{"type":"item.completed","item":{"id":"item_0","type":"reasoning","text":"**Confirming exact output**"}}
+{"type":"item.completed","item":{"id":"item_0","type":"reasoning","text":"**Requesting exact output**"}}
 {"type":"item.completed","item":{"id":"item_1","type":"agent_message","text":"5"}}
-{"type":"turn.completed","usage":{"input_tokens":9238,"cached_input_tokens":6656,"output_tokens":20}}
+{"type":"turn.completed","usage":{"input_tokens":9128,"cached_input_tokens":6912,"output_tokens":21}}
