+++ src/lib/targets/builtins/codex/e2e/expected/output-stream-json.trace.json
@@ -5,6 +5,8 @@
     "--sandbox",
     "workspace-write",
     "--json",
+    "-m",
+    "gpt-5.1-codex-mini",
     "--disable",
     "web_search_request",
     "-c",
