+++ src/lib/targets/builtins/codex/e2e/expected/passthrough.stdout.txt
@@ -1,1 +1,1 @@
-codex-cli-exec 0.92.0
+codex-cli-exec 0.89.0
