+++ src/lib/targets/builtins/codex/e2e/expected/sandbox-workspace-write.trace.json
@@ -4,6 +4,8 @@
     "exec",
     "--sandbox",
     "workspace-write",
+    "-m",
+    "gpt-5.1-codex-mini",
     "--disable",
     "web_search_request",
     "-c",
