+++ src/lib/targets/builtins/codex/e2e/expected/web-on.stderr.txt
@@ -1,21 +1,21 @@
-OpenAI Codex v0.92.0 (research preview)
+OpenAI Codex v0.89.0 (research preview)
 --------
 workdir: /Users/joeroddy/Documents/dev/projects/open-source/omniagent
-model: gpt-5.2-codex
+model: gpt-5.1-codex-mini
 provider: openai
 approval: never
 sandbox: workspace-write [workdir, /tmp, $TMPDIR]
 reasoning effort: high
 reasoning summaries: auto
-session id: 019c0af3-9378-74f3-a799-b14f84901ed9
+session id: 019c14bd-f537-7c20-8040-224839764fde
 --------
 user
 Output exactly: 5
 mcp startup: no servers
 
 thinking
-**Enforcing fixed output count**
+**Confirming output requirement**
 codex
 5
 tokens used
-2,602
+2,249
